+++ src/ltm/bn_mp_set_long_long.c
@@ -1,4 +1,4 @@
-#include <tommath_private.h>
+#include "tommath_private.h"
 #ifdef BN_MP_SET_LONG_LONG_C
 /* LibTomMath, multiple-precision integer library -- Tom St Denis
  *
@@ -11,12 +11,10 @@
  *
  * The library is free for all purposes without any express
  * guarantee it works.
- *
- * Tom St Denis, tstdenis82@gmail.com, http://libtom.org
  */
 
 /* set a platform dependent unsigned long long int */
-MP_SET_XLONG(mp_set_long_long, uint64_t)
+MP_SET_XLONG(mp_set_long_long, unsigned long long)
 #endif
 
 /* ref:         $Format:%D$ */
